Push raw events also through eventstream, but flag them to not be consumed
This "raw" feature is only used by `create_sample_event`, and is meant
to be used to bypass notifications and other post processing events.

Prior to this patch, this meant that all sample events were skipping the
eventstream entirely. This is a concern for both production (since you
can trigger sample events through the UI) and local for load-mocks,
which bypass being stored into the eventstream.